+++ src/app/(main)/grouping/[groupId]/components/detail/MemberSection.client.tsx
@@ -29,7 +29,7 @@ export default function MemberSection() {
           onClick={() => router.push(`${pathname}/members`)}
         >
           <p>전체 보기</p>
-          <Image src="/icons/24/navigate-next.svg" alt="navigate-next" width={24} height={24} />
+          <Image src="/icons/24/navigate_next.svg" alt="navigate_next" width={24} height={24} />
         </div>
       </div>
       <Spacing size={20} />
