+++ src/app/(main)/grouping/[groupId]/members/components/MemberList.client.tsx
@@ -42,7 +42,7 @@ function MemberItem({ member }: MemberItemProps) {
           </Flex>
         </div>
         <IconButton>
-          <Image src="/icons/24/navigate-next.svg" alt="navigate-next" width={24} height={24} />
+          <Image src="/icons/24/navigate_next.svg" alt="navigate_next" width={24} height={24} />
         </IconButton>
       </Flex>
     </Link>
