+++ src/app/globals.css
@@ -45,99 +45,57 @@ textarea::-webkit-scrollbar-thumb {
   height: 1rem;
 }
 
-html {
-  font-size: 4vw;
-}
-
-@media (min-width: 420px) {
-  html {
-    font-size: 16px;
-  }
-}
-
 .react-datepicker {
-  font-family: 'Noto Sans KR', serif;
-  border: none !important;
-  border-radius: 0;
-  width: 20.125rem;
-  height: 17.5rem;
-}
-
-.react-datepicker__week {
-  width: 100%;
+  @apply w-full !border-none;
 }
 
 .react-datepicker__month-container {
-  width: 100%;
+  @apply w-full;
 }
 
 .react-datepicker .react-datepicker__header {
-  color: #3a3a3a;
-  background-color: white;
-  border-bottom: none;
-  border-radius: 0;
-  width: 100%;
+  @apply flex flex-col gap-16 border-none bg-inherit pb-28 pt-0;
 }
 
 .react-datepicker .react-datepicker__month {
-  width: 100%;
-  margin: 0;
+  @apply m-0 flex flex-col gap-20;
 }
 
 .react-datepicker .react-datepicker__day--outside-month {
-  cursor: default;
-  visibility: hidden;
+  @apply invisible cursor-default;
 }
+
 .react-datepicker .react-datepicker__header .react-datepicker__day-names {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 40px;
-}
-
-.react-datepicker .react-datepicker__month .react-datepicker__week {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.react-datepicker .react-datepicker__day {
-  color: #aaaaaa;
-  font-weight: 700;
-  font-size: 0.75rem;
+  @apply flex justify-between;
 }
 
 .react-datepicker .react-datepicker__day-name {
-  color: #3a3a3a;
-  font-weight: 700;
-  font-size: 0.75rem;
+  @apply m-0 flex h-32 w-32 items-center justify-center text-subtitle-3 text-sign-tertiary;
+}
+
+.react-datepicker .react-datepicker__month .react-datepicker__week {
+  @apply flex justify-between;
+}
+
+.react-datepicker .react-datepicker__day {
+  @apply flex h-32 w-32 items-center justify-center text-paragraph-1 text-sign-secondary;
 }
 
 .react-datepicker .react-datepicker__day:hover {
-  color: #aaaaaa;
-  font-weight: 700;
-  font-size: 0.75rem;
-  border-radius: 100%;
+  @apply flex h-32 w-32 items-center justify-center rounded-full bg-primary-light text-white;
 }
 
-.react-datepicker .react-datepicker__day--selected {
-  color: #ffffff;
-  font-weight: 700;
-  font-size: 0.75rem;
-  border-radius: 100%;
-  background-color: #1249fc !important;
+.react-datepicker .react-datepicker__day--selected,
+.react-datepicker .react-datepicker__day--keyboard-selected {
+  @apply flex h-32 w-32 items-center justify-center rounded-full bg-primary text-white;
 }
 
 .react-datepicker .react-datepicker__day--selected:hover {
-  color: #ffffff;
-  font-weight: 700;
-  font-size: 0.75rem;
-  border-radius: 100%;
-  background-color: #1249fc !important;
+  @apply flex h-32 w-32 items-center justify-center rounded-full bg-primary-light text-white;
 }
 
 .react-datepicker .react-datepicker__triangle {
-  display: none;
+  @apply hidden;
 }
 
 .swiper-slide {
