+++ src/components/Calendar/index.ts
@@ -1,0 +1,1 @@
+export { default as Calendar } from './Calendar';
